AMD: Move common stuff from Romecrb to amd/common/boot, part 2.
