+++ src/mainboard/amd/romecrb/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 
 [dependencies]
 arch = { path = "../../../arch/x86/x86_64"}
+boot = { path = "../../../soc/amd/common/boot" }
 clock = { path = "../../../drivers/clock"}
 console = { path = "../../../console" }
 cpu = { path = "../../../cpu/amd" }
